LayoutTests/ietestcenter/Javascript/15.10.7.3-1-expected.txt

15.10.7.3-1 (TestCases/chapter15/15.10/15.10.7/15.10.7.3/15.10.7.3-1.js) RegExp.prototype.ignoreCase is of type Boolean

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
